Upgraded Laravel 5.8 and created models

+++ KeyPiryWeb/database/migrations/2018_11_27_105724_create_products_table.php
@@ -16,8 +16,8 @@ class CreateProductsTable extends Migration
     {
         //Schema::enableForeignKeyConstraints();
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('category_id')->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('category_id')->nullable();
             $table->string('name',191)->nullable();
             $table->text('description')->nullable();
             $table->string('image',191)->nullable();
