+++ KeyPiryWeb/database/migrations/2018_11_28_124110_create_subproducts_images_table.php
@@ -14,9 +14,9 @@ class CreateSubproductsImagesTable extends Migration
     public function up()
     {
         Schema::create('subproducts_images', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('image')->nullable();
-            $table->unsignedInteger('subproduct_id')->nullable();
+            $table->unsignedBigInteger('subproduct_id')->nullable();
             $table->timestamps();
 
             $table->foreign('subproduct_id')
